+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+## [0.3.0] - 13.06.2024
+
+### Added
+- `thunk` extension ([#6](https://github.com/purplenoodlesoop/pure/issues/6)).
+- `TaggedRecord` base class.
+- A list of primitive functions.
+
+### Changed
+- Required Dart SDK version is now `3.0.0`.
+- Updated Recursion trampolines to use Sealed Classes.
+- Make `constant` extension more flexible ([#5](https://github.com/purplenoodlesoop/pure/issues/5))
+
+### Fixed
+- Incorrect hash calculation ([#8](https://github.com/purplenoodlesoop/pure/commit/4e0e1aad88743ff9a651f12a5d642df38a9d7e83)).
+
+### Removed
+- [BREAKING] Convenience import mixin.
+
 ## [0.2.0] - 20.03.2022
 
 ### Changed